ext4l: Fix format warning in mballoc.c
Fix format warning: atomic64_t.counter is 'long' on 32-bit systems,
so cast to 'long long' for %lld format.

Co-developed-by: Claude <noreply@anthropic.com>

+++ fs/ext4l/mballoc.c
@@ -3907,9 +3907,9 @@ void ext4_mb_release(struct super_block *sb)
 				atomic_read(&sbi->s_bal_breaks),
 				atomic_read(&sbi->s_mb_lost_chunks));
 		ext4_msg(sb, KERN_INFO,
-		       "mballoc: %u generated and it took %llu",
+		       "mballoc: %u generated and it took %lld",
 				atomic_read(&sbi->s_mb_buddies_generated),
-				atomic64_read(&sbi->s_mb_generation_time));
+				(long long)atomic64_read(&sbi->s_mb_generation_time));
 		ext4_msg(sb, KERN_INFO,
 		       "mballoc: %u preallocated, %u discarded",
 				atomic_read(&sbi->s_mb_preallocated),
